requirements.txt
Django==1.8.2
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Wand==0.3.8
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
autoflake==0.6.6
billiard==3.3.0.19
boto==2.34.0
cassandra-driver==2.1.3
celery==3.1.17
cqlengine==0.20.0
django-appconf==0.6
django-autocomplete-light==2.0.3
django-celery==3.1.16
django-commando==0.1.3
django-elasticache==0.0.2
django-embed-video==0.11
django-feedly==0.6.4
django-grunt==1.1.0
django-multiupload==0.3
django-pagination==1.0.7
django-pipeline==1.4.6
django-social-auth==0.7.28
django-static-precompiler==0.8
django-storages==1.1.8
django-test-without-migrations==0.2
futures==2.2.0
html2text==2014.12.5
httplib2==0.9.1
icalendar==3.9.0
kombu==3.0.24
mutagen==1.27
oauth2==1.5.211
psycopg2==2.5.4
pyflakes==0.9.0
pygeoip==0.3.2
pylibmc==1.4.1
pysolr==3.2.0
python-dateutil==2.3
python-memcached==1.53
python-openid==2.2.5
pytz==2014.10
redis==2.10.3
requests==2.5.0
simplejson==3.6.5
six==1.9.0
sorl-thumbnail==12.2
stream-framework==1.1.1
wsgiref==0.1.2
